Move createTimestampFile and walkDirectory to helpers-pure.ts
