+++ extensions/ql-vscode/src/codeql-cli/cli.ts
@@ -21,12 +21,12 @@ import {
   assertNever,
   getErrorMessage,
   getErrorStack,
+  walkDirectory,
 } from "../pure/helpers-pure";
 import { QueryMetadata, SortDirection } from "../pure/interface-types";
 import { BaseLogger, Logger, ProgressReporter } from "../common";
 import { CompilationMessage } from "../pure/legacy-messages";
 import { sarifParser } from "../common/sarif-parser";
-import { walkDirectory } from "../helpers";
 import { App } from "../common/app";
 import { QueryLanguage } from "../common/query-language";
