+++ extensions/ql-vscode/src/pure/helpers-pure.ts
@@ -5,7 +5,9 @@
  * Helper functions that don't depend on vscode or the CLI and therefore can be used by the front-end and pure unit tests.
  */
 
+import { join } from "path";
 import { RedactableError } from "./errors";
+import { ensureDir, opendir, writeFile } from "fs-extra";
 
 /**
  * This error is used to indicate a runtime failure of an exhaustivity check enforced at compile time.
@@ -66,4 +68,40 @@ export function asError(e: unknown): Error {
   }
 
   return e instanceof Error ? e : new Error(String(e));
+}
+
+/**
+ * Creates a file in the query directory that indicates when this query was created.
+ * This is important for keeping track of when queries should be removed.
+ *
+ * @param queryPath The directory that will contain all files relevant to a query result.
+ * It does not need to exist.
+ */
+export async function createTimestampFile(storagePath: string) {
+  const timestampPath = join(storagePath, "timestamp");
+  await ensureDir(storagePath);
+  await writeFile(timestampPath, Date.now().toString(), "utf8");
+}
+
+/**
+ * Recursively walk a directory and return the full path to all files found.
+ * Symbolic links are ignored.
+ *
+ * @param dir the directory to walk
+ *
+ * @return An iterator of the full path to all files recursively found in the directory.
+ */
+export async function* walkDirectory(
+  dir: string,
+): AsyncIterableIterator<string> {
+  const seenFiles = new Set<string>();
+  for await (const d of await opendir(dir)) {
+    const entry = join(dir, d.name);
+    seenFiles.add(entry);
+    if (d.isDirectory()) {
+      yield* walkDirectory(entry);
+    } else if (d.isFile()) {
+      yield entry;
+    }
+  }
 }
