+++ extensions/ql-vscode/src/run-queries-shared.ts
@@ -2,7 +2,7 @@ import * as messages from "./pure/messages-shared";
 import * as legacyMessages from "./pure/legacy-messages";
 import { DatabaseInfo, QueryMetadata } from "./pure/interface-types";
 import { join, parse, dirname, basename } from "path";
-import { createTimestampFile, showAndLogWarningMessage } from "./helpers";
+import { showAndLogWarningMessage } from "./helpers";
 import {
   ConfigurationTarget,
   Range,
@@ -29,7 +29,7 @@ import { DatabaseManager } from "./databases/local-databases";
 import { DecodedBqrsChunk, EntityValue } from "./pure/bqrs-cli-types";
 import { BaseLogger, extLogger } from "./common";
 import { generateSummarySymbolsFile } from "./log-insights/summary-parser";
-import { getErrorMessage } from "./pure/helpers-pure";
+import { createTimestampFile, getErrorMessage } from "./pure/helpers-pure";
 
 /**
  * run-queries.ts
